requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
comtypes==1.1.7
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2

MarkupSafe==1.1.1

pathlib==1.0.1
pdf2image==1.14.0
Pillow==8.1.0
pycryptodome==3.10.1

pypiwin32==223
pytesseract==0.3.7

pywin32==228
requests==2.25.1
six==1.15.0
tenacity==5.1.5
urllib3==1.26.3
Werkzeug==1.0.1
